requirements.txt

botbuilder-core==4.17.0
botbuilder-schema==4.17.0
botbuilder-integration-aiohttp==4.17.0    

aiohttp          
yarl==1.9.4               
gunicorn==21.2.0


openai==1.6.1
tiktoken==0.5.2


azure-identity==1.15.0
msal==1.31.1
pyjwt==2.8.0


python-dotenv==1.0.0
requests==2.31.0
certifi==2023.11.17
charset-normalizer==3.3.2
